shade-finder-api/shade-finder-fast-api-service/app/requirements.txt
face_detection_tflite==0.6.0
imutils==0.5.4
matplotlib==3.9.0
numpy==1.26.4
opencv-python-headless==4.10.0.82
pandas==2.2.2
Pillow==10.3.0
scikit_learn==1.5.0
skin_tone_classifier==1.2.4
fastapi==0.111.0
fastapi-cli==0.0.4
uvicorn==0.30.1